refactor: ai.ts scripts로 이동
